Fix: Move fixtures into dedicated namespace
